keep inlineChat.contribution purely registrations of stuff
